requirements.txt
decouple
boto3==1.9.89
botocore==1.12.89
dj-database-url==0.4.2
dj-static==0.0.6
django-cors-headers==2.2.0
django-storages==1.7.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Pillow==5.2.0
psycopg2-binary
PyJWT==1.6.4
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.5
whitenoise==3.3.1

